// app/types/jobTypes.ts
export interface JobDetails {
    description: string;
    responsibilities: string;
    idealCandidate: string;
    whenAndWhere: string;
    datePosted: string;
    deadline: string;
    location: string[];
    startDate: string;
    endDate: string;
    categories: string[];
    requiredSkills: string[];
  }
  